Create new BottomSection component
